Add Bundle Creation Logic (#1)
* chore: add .gitignore

* build: add project skeleton using poetry

* feat(bundle): add the BundleConfig dataclass

Add the BundleConfig dataclass, which contains directives and other metadata associated
with a Bundle. A BundleConfig is written in TOML on disk. Currently, a BundleConfig has
the following information:

* A name.
* An optional docker command to execute in an associated container.
* An optional docker entrypoint to execute in an associated container.
* An optional dictionary container volumes to be mounted at runtime.

Additionally, add methods to serialize and deserialize a BundleConfig to/from TOML
strings and files.

Add associated tests and documentation.

* feat(bundle): add logic to create Bundle files

Add the first pass at the logic needed to create Bundle files. Specifically,
add the `BundleInfo` class, which contains metadata about a Bundle; and implement
the `create_bundle` function, which creates a Bundle from a specified directory with
a yardgoat.toml file. Lastly, add tests, and expand documentation.