feat(engine): implement docker yardgoat runtime (#5)
* feat(bundle): add utility functions for extracting bundle data

Add two utility functions: `extract` and `extract_volume_files`. The function `extract` extracts
a Bundle file to a specified directory, and the function `extract_volume_files` extracts all files
listed under the `[volumes]` heading in the yardgoat.toml file. The latter function is useful for when
one needs to execute a docker container and mount volumes/files.

* feat(engine): add the Engine protocol and DockerEngine implementation

Add the `Engine` protocol, which defines an interface for Bundle building and execution.
Additionally, implement the class `DockerEngine`, which implements the Engine protocol,
and enables yardgoat to execute Docker-based batch jobs. Beyond the `Engine` protocol, define
the `Artifact` dataclass, which represents an assembled Bundle for a given `Engine` runtime.
Lastly, create the factory function `new`, which is exposed at the top level of the engine subpackage.
This function allows users to specify the runtime they want via a string or enum, and get the concrete
`Engine` implementation without having them manually create it themselves.