feat(worker): implement the Worker class and core logic (#6)
* feat(worker): add preliminary types and exceptions

Add the following dataclasses: ObjectURI, which represents the URI of a yardgoat object;
and Job, which represents a yardgoat job. Additionally, add the JobQueue protocol, which
defines an interface for submitting and receiving Jobs via some type of queue. Lastly,
add the AlreadyCompletedError.

* feat(worker): add an AWS SQS job queue implementation

Create a JobQueue implementation that uses AWS SQS as its queueing system. Implement
supporting tests.

* docs(worker): add missing docstrings for worker subpackage

* test(worker): add timeout test for SQSJobQueue.get_open_request

* feat(worker): implement the Worker class

Add the Worker class, which combines the entire Yardgoat lifecycle logic for executing jobs.
Specifically, it will:

1. Pull down the next available job
2. Download the associated Bundle
3. Build the Bundle
4. Execute the built Artifact

Further, a worker is able to repeat this logic repeatedly via the `loop` method.

Add associated tests.